requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==3.0.10
Flask-Migrate==4.0.4
python-dotenv==1.0.0
PyMySQL==1.0.3
requests==2.31.0
google-generativeai==0.8.2
Flask-JWT-Extended==4.4.4
cryptography